Create Music Player, Semi-Finished
In the older version of this project, I had to use discord notifications and read them using Tesseract. This created a few problems, like that it couldn't read URLs very well. So in order to fix this, I got rid of tesseract and am now only dealing with .txt file and .rtf files that are being updated constantly through iCloud.